+++ data-browser/src/components/SideBar.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import * as React from 'react';
-import { useArray, useResource, useStore, useTitle } from '@tomic/react';
-import { properties, ResourceStatus } from '@tomic/lib';
+import { useArray, useResource, useTitle } from '@tomic/react';
+import { properties } from '@tomic/lib';
 import { useHover } from '../helpers/useHover';
 import { useSettings } from '../helpers/AppSettings';
 import { useWindowSize } from '../helpers/useWindowSize';
@@ -206,7 +206,7 @@ function SideBarDrive({ handleClickItem }: SideBarDriveProps): JSX.Element {
             />
           );
         })
-      ) : drive.getStatus() == ResourceStatus.loading ? null : (
+      ) : drive.loading ? null : (
         <SideBarErr>
           {drive.getError()?.message || 'Could not load this baseURL'}
         </SideBarErr>