+++ lib/src/parse.ts
@@ -1,10 +1,13 @@
-import { Resource, ResourceStatus, unknownSubject } from './resource';
+import { isArray, JSONValue, Store } from '.';
+import { Resource, unknownSubject } from './resource';
 import { JSONObject } from './value';
 
 /** Parses an JSON-AD object containing a resoure, adds it to the input Resource */
 export function parseJsonADResource(
   jsonObject: JSONObject,
   resource: Resource,
+  /** Pass a Store if you want to add the parsed resources to it */
+  store?: Store,
 ): void {
   try {
     for (const key in jsonObject) {
@@ -24,22 +27,75 @@ export function parseJsonADResource(
         resource.setSubject(subject);
         continue;
       }
+      const value = jsonObject[key];
       try {
-        resource.setUnsafe(key, jsonObject[key]);
+        // Resource values can be either strings (URLs) or full Resources, which in turn can be either Anonymous (no @id) or Named (with an @id)
+        if (isArray(value)) {
+          const newarr = value.map(val =>
+            parseJsonAdResourceValue(store, val, resource, key),
+          );
+          resource.setUnsafe(key, newarr);
+        } else if (typeof value === 'string') {
+          resource.setUnsafe(key, value);
+        } else if (typeof value === 'number') {
+          resource.setUnsafe(key, value);
+        } else if (typeof value === 'boolean') {
+          resource.setUnsafe(key, value);
+        } else {
+          const subject = parseJsonAdResourceValue(store, value, resource, key);
+          resource.setUnsafe(key, subject);
+        }
       } catch (e) {
         throw new Error(
-          `Failed creating value for key ${key} in resource ${resource.getSubject()}. ${e.message
+          `Failed creating value ${value} for key ${key} in resource ${resource.getSubject()}. ${e.message
           }`,
         );
       }
     }
-    resource.setStatus(ResourceStatus.ready);
+    resource.loading == false;
+    store && store.addResource(resource);
   } catch (e) {
     e.message = 'Failed parsing JSON ' + e.message;
     resource.setError(e);
+    resource.loading == false;
+    store && store.addResource(resource);
     throw e;
   }
   return;
+}
+
+type StringOrNestedResource = string | JSONObject;
+
+/**
+ * Parses a JSON-AD Value. If it's a string, it takes its URL. If it's an
+ * Object, it will parse it as a Resource. It will add the string property to
+ * the Resource.
+ */
+function parseJsonAdResourceValue(
+  store: Store,
+  value: JSONValue,
+  resource: Resource,
+  key: string,
+): StringOrNestedResource {
+  if (typeof value === 'string') {
+    return value;
+  }
+  if (value.constructor === {}.constructor) {
+    if (Object.keys(value).includes('@id')) {
+      // It's a named resource that needs to be put in the store
+      const nestedSubject = value['@id'];
+      const nestedResource = new Resource(nestedSubject);
+      parseJsonADResource(value as JSONObject, nestedResource, store);
+      if (store) {
+        store.addResource(nestedResource);
+      }
+      return nestedSubject;
+    } else {
+      // It's an anonymous nested Resource
+      return value as JSONObject;
+    }
+  }
+  throw new Error(`Value ${value} in ${key} not a string or a nested Resource`);
 }
 
 /** Parsees a JSON-AD array string, returns array of Resources */